entity article

+++ src/Entity/CategoryArticle.php
@@ -14,20 +14,21 @@ class CategoryArticle
 {
     #[ORM\Id]
     #[ORM\GeneratedValue]
-    #[ORM\Column]
+    #[ORM\Column(type: 'integer')]
     private ?int $id = null;
 
     #[ORM\Column(length: 100, nullable: true)]
     private ?string $name = null;
 
-    #[ORM\ManyToOne(inversedBy: 'service')]
-    private ?Article $articles = null;
+    #[ORM\ManyToMany(targetEntity: Article::class, inversedBy: 'categoryArticles')]
+    private Collection $articles;
 
     #[ORM\ManyToMany(targetEntity: Service::class, inversedBy: 'categoryArticle')]
     private Collection $services;
 
     public function __construct()
     {
+        $this->articles = new ArrayCollection();
         $this->services = new ArrayCollection();
     }
 
@@ -41,25 +42,36 @@ class CategoryArticle
         return $this->name;
     }
 
-    public function setName(?string $name): static
+    public function setName(?string $name): self
     {
         $this->name = $name;
 
         return $this;
     }
 
-    public function getArticles(): ?Article
+    /**
+     * @return Collection<int, Article>
+     */
+    public function getArticles(): Collection
     {
         return $this->articles;
     }
 
-    public function setArticles(?Article $articles): static
+    public function addArticle(Article $article): self
     {
-        $this->articles = $articles;
+        if (!$this->articles->contains($article)) {
+            $this->articles[] = $article;
+        }
 
         return $this;
     }
 
+    public function removeArticle(Article $article): self
+    {
+        $this->articles->removeElement($article);
+
+        return $this;
+    }
     /**
      * @return Collection<int, Service>
      */
@@ -68,20 +80,22 @@ class CategoryArticle
         return $this->services;
     }
 
-    public function addService(Service $service): static
+    public function addService(Service $service): self
     {
         if (!$this->services->contains($service)) {
-            $this->services->add($service);
-            $service->addCategoryArticle($this);
+            $this->services[] = $service;
+            $service->addCategoryArticle($this); // Cette ligne pourrait ne pas être nécessaire selon la logique de votre application
         }
 
         return $this;
     }
 
-    public function removeService(Service $service): static
+    public function removeService(Service $service): self
     {
-        $this->services->removeElement($service);
-        $service->removeCategoryArticle($this);
+        if ($this->services->removeElement($service)) {
+            // Optionnel: Si vous maintenez la relation des deux côtés
+            $service->removeCategoryArticle($this);
+        }
 
         return $this;
     }
